fix some 64-bit warnings; tweak warnings flags in premake

+++ src/uia/DocumentProvider.cpp
@@ -170,7 +170,7 @@ HRESULT STDMETHODCALLTYPE SumatraUIAutomationDocumentProvider::GetRuntimeId(SAFE
         return E_OUTOFMEMORY;
 
     // RuntimeID magic, use hwnd to differentiate providers of different windows
-    int rId[] = { (int)canvasHwnd, SUMATRA_UIA_DOCUMENT_RUNTIME_ID };
+    LONG rId[] = { (LONG)canvasHwnd, SUMATRA_UIA_DOCUMENT_RUNTIME_ID };
     for (LONG i = 0; i < 2; i++) {
         HRESULT hr = SafeArrayPutElement(psa, &i, (void*)&(rId[i]));
         CrashIf(FAILED(hr));
